+++ test_ak_ak_code.py
@@ -1,4 +1,0 @@
-import akshare as ak
-import inspect
-
-print(inspect.getsource(ak.stock_zh_a_hist_min_em))